Day 2, part 2
Not too bad. I had to think about how to get "our" results but other
than that little changed. It would have been better in part one to move
the hash-ref for our-full-throw to its own function. If that were the
case this part could have been reduced to one new mapping and one
changed function.