Day 3, part 1 and 2
Had to write my own group function, which seems but I couldn't find one
in racket.
There may also be unnecessary work in dividing the list in half by
making it a vector briefly. I probably could have used the group
function from part 2 if I knew I'd have to write it (or if racket's was
easier to find).