Day 7, Part 2
Had to refactor a little of part 1 so I left it as a separate commit.
Using findf was a new twist for me instead of using continuations. It
aborts evaluation of the list once the procedure/function passed in
returns true. I got to play with it as below:

(define test-or-map '((+ 1 2) (+ 3 4) (error (quote too-far)) (+ 5 6)))
; testing the error:
(map (lambda (b) (eval b)) test-or-map);; results in the error(!)
; making sure findf aborts when an answer is found:
(findf (lambda (b) (> (eval b) 5)) test-or-map);; results in '(+ 3 4)